refactor: Move SuperJson tests to proper files
