+++ src/schema-tools/superjson/index.ts
@@ -1,3 +1,3 @@
 export * from './mutate';
-export * from './schema';
+export * from './normalize';
 export * from './utils';
